temporarily move userfunctions code to script plugin
