+++ packages/apps/plugins/plugin-script/package.json
@@ -9,16 +9,32 @@
   "exports": {
     ".": {
       "browser": "./dist/lib/browser/index.mjs",
+      "node": {
+        "default": "./dist/lib/node/index.cjs"
+      },
       "types": "./dist/types/src/index.d.ts"
+    },
+    "./edge": {
+      "browser": "./dist/lib/browser/edge/index.mjs",
+      "node": {
+        "default": "./dist/lib/node/edge/index.cjs"
+      },
+      "types": "./dist/types/src/edge/index.d.ts"
     },
     "./meta": {
       "browser": "./dist/lib/browser/meta.mjs",
+      "node": {
+        "default": "./dist/lib/node/meta.cjs"
+      },
       "types": "./dist/types/src/meta.d.ts"
     }
   },
   "types": "dist/types/src/index.d.ts",
   "typesVersions": {
     "*": {
+      "edge": [
+        "dist/types/src/edge/index.d.ts"
+      ],
       "meta": [
         "dist/types/src/meta.d.ts"
       ]
@@ -56,6 +72,7 @@
     "@preact/signals-core": "^1.6.0",
     "codemirror": "^6.0.1",
     "esbuild-wasm": "^0.16.14",
+    "iso-did": "^1.6.0",
     "lodash.get": "^4.4.2"
   },
   "devDependencies": {
