+++ packages/apps/plugins/plugin-script/src/edge/codec.ts
