+++ packages/apps/plugins/plugin-script/src/edge/header-codec.ts
