+++ packages/apps/plugins/plugin-script/src/edge/hub-protocol.ts
@@ -4,9 +4,8 @@
 
 import { type DID } from 'iso-did/types';
 
+import { type Credential } from '@dxos/client/halo';
 import { type PublicKey } from '@dxos/keys';
-
-import { type Credential } from '../../halo';
 
 // TODO(burdon): Factor out to @dxos/hub-protocol.
