+++ packages/devtools/cli/package.json
@@ -31,6 +31,7 @@
     "readme": "oclif readme --multi"
   },
   "dependencies": {
+    "@braneframe/plugin-script": "workspace:*",
     "@braneframe/types": "workspace:*",
     "@dxos/agent": "workspace:*",
     "@dxos/async": "workspace:*",
