+++ packages/devtools/cli/tsconfig.json
@@ -23,6 +23,9 @@
   ],
   "references": [
     {
+      "path": "../../apps/plugins/plugin-script"
+    },
+    {
       "path": "../../apps/templates/bare-template"
     },
     {
