+++ packages/sdk/client/src/services/index.ts
@@ -15,10 +15,3 @@ export { SharedWorkerConnection } from './shared-worker-connection';
 export { fromWorker, WorkerClientServices } from './worker-client-services';
 export { type AgentHostingProviderClient, AgentManagerClient } from './agent-hosting-provider';
 export { FakeAgentHostingProvider } from './fake-agent-hosting-provider';
-export {
-  uploadWorkerFunction,
-  getUserFunctionIdInMetadata,
-  setUserFunctionIdInMetadata,
-  type UserFunctionUploadResult,
-  type UploadWorkerProps,
-} from './userfunctions';
