+++ tsconfig.paths.json
@@ -176,6 +176,9 @@
       "@braneframe/plugin-registry/meta": [
         "packages/apps/plugins/plugin-registry/src/meta.ts"
       ],
+      "@braneframe/plugin-script/edge": [
+        "packages/apps/plugins/plugin-script/src/edge/index.ts"
+      ],
       "@braneframe/plugin-script": [
         "packages/apps/plugins/plugin-script/src/index.ts"
       ],
